+++ CHANGES.rst
@@ -5,6 +5,7 @@ Changelog
 ------------------
 
 - #2344 Fix error on sample copy when fields are hidden
+- #2334 Remove legacy reports
 - #2341 Handle nonexisting UIDs gracefully in referencewidget
 - #2340 Fix UID copy method in AT->DX content migrator
 - #2332 Fix unauthorized error when accessing immediate results entry view with a client contact user